Added maps to pull and graphing notebooks
Note, on the map that was already in the Jobs data pull, I was unable to run it on my computer. It may be good to make sure your directory is clear before running it because I think it assumes I have access to a jobs_map_data function.

Additonally, I have added increased functionality to the graphing maps function. A user can now specify their color palette as an input and the unit of measurement (dollars, percent, or none).